Add Popup and DeleteConfirmation to ShowAttendance, refactor DeleteConfirmation into reusable component, add title prop to ShowDeliverable instance
