+++ README.md
@@ -10,6 +10,8 @@ https://auth0.com/docs/quickstart/spa/react/interactive
 • Student tag should go inside submission materials on grading page
   display student who is being graded alongside the link
 
+• JSX for attendance notes has been removed from header. Do we need it elswhere?
+
 • figure out redirect on reload GradeStudentDeliverable + ShowPerson
   useRedirectOnRefresh
