+++ src/components/DeleteConfirmation/DeleteConfirmation.css
@@ -1,0 +1,42 @@
+.confirmation {
+  top: 50%;
+  left: 50%;
+  z-index: 2;
+  padding: 24px;
+  display: flex;
+  min-width: 380px;
+  position: absolute;
+  border-radius: 5px;
+  flex-direction: column;
+  transform: translate(-50%, -50%);
+  border: 1px solid var(--border-color);
+  background-color: var(--content-background);
+  box-shadow: 0px 5px 5px 5px rgba(0, 0, 0, 0.75);
+  -moz-box-shadow: 0px 5px 5px 5px rgba(0, 0, 0, 0.75);
+  -webkit-box-shadow: 0px 5px 5px 5px rgba(0, 0, 0, 0.75);
+}
+
+.confirmation * {
+  margin: 0;
+  line-height: 1;
+}
+
+.confirmation h1 {
+  margin-bottom: 14px;
+}
+
+.confirmation p {
+  font-weight: 300;
+  text-align: center;
+  letter-spacing: .75px;
+}
+
+.confirmation button {
+  width: 80px;
+  height: 24px;
+  font-size: 12px;
+  font-weight: 600;
+  border-radius: 5px;
+  margin: 24px 8px 0px;
+  padding: 6px 0px 18px;
+}
